+++ packages/core/index.ts
@@ -2,7 +2,6 @@ export * from "./storage.js";
 export * from "./util.js";
 export * from "./moments.js";
 export * from "./timeline.js";
-export * from "./extendable.js";
 
 /**
  * Behavior exports
